Make ResourceName an interface and create AbstractResourceName as a drop-in replacement.
Having ResourceName being an interface makes it easier to create classes for resources which are not file system based. For instance, on Android you might refer to a "raw resource" which is identified by an integer in Java code.
